Refactoring structure Main
